# Add additional parameters to Schema for validation and annotation