--- a/pydantic/__init__.py
+++ b/pydantic/__init__.py
@@ -2,9 +2,10 @@
 from .env_settings import BaseSettings
 from .error_wrappers import ValidationError
 from .errors import *
-from .fields import Required, Schema
+from .fields import Required
 from .main import BaseConfig, BaseModel, create_model, validate_model, validator
 from .parse import Protocol
 from .types import *
 from .version import VERSION
 from . import dataclasses
+from .schema import Schema