--- a/pydantic/fields.py
+++ b/pydantic/fields.py
@@ -1,10 +1,25 @@
 import inspect
+from decimal import Decimal
 from enum import IntEnum
 from typing import Any, Callable, List, Mapping, NamedTuple, Pattern, Set, Tuple, Type, Union
 
 from . import errors as errors_
 from .error_wrappers import ErrorWrapper
-from .types import Json, JsonWrapper
+from .types import (
+    DSN,
+    ConstrainedDecimal,
+    ConstrainedFloat,
+    ConstrainedInt,
+    ConstrainedStr,
+    EmailStr,
+    Json,
+    JsonWrapper,
+    UrlStr,
+    condecimal,
+    confloat,
+    conint,
+    constr,
+)
 from .utils import display_as_type, list_like
 from .validators import NoneType, dict_validator, find_validators, not_none_validator
 
@@ -36,16 +51,98 @@ class Validator(NamedTuple):
 
 class Schema:
     """
-    Used to provide extra information about a field in a model schema.
+    Used to provide extra information about a field in a model schema. The parameters will be
     """
 
-    __slots__ = 'default', 'alias', 'title', 'extra'
+    __slots__ = (
+        'default',
+        'alias',
+        'title',
+        'description',
+        'gt',
+        'ge',
+        'lt',
+        'le',
+        'min_length',
+        'max_length',
+        'regex',
+        'extra',
+    )
 
-    def __init__(self, default, *, alias=None, title=None, **extra):
+    def __init__(
+        self,
+        default,
+        *,
+        alias: str = None,
+        title: str = None,
+        description: str = None,
+        gt: float = None,
+        ge: float = None,
+        lt: float = None,
+        le: float = None,
+        min_length: int = None,
+        max_length: int = None,
+        regex: str = None,
+        **extra,
+    ):
         self.default = default
         self.alias = alias
         self.title = title
+        self.description = description
         self.extra = extra
+        self.gt = gt
+        self.ge = ge
+        self.lt = lt
+        self.le = le
+        self.min_length = min_length
+        self.max_length = max_length
+        self.regex = regex
+
+
+# base classes, classes blacklist, args to check and assign
+_numeric_types = (int, float, Decimal)
+_blacklist = (EmailStr, DSN, UrlStr, ConstrainedStr, ConstrainedInt, ConstrainedFloat, ConstrainedDecimal)
+_str_attrs = ('max_length', 'min_length', 'regex')
+_numeric_attrs = ('gt', 'lt', 'ge', 'le')
+_map_types_const = {str: constr, int: conint, float: confloat, Decimal: condecimal}
+
+
+def get_annotation_from_schema(annotation, schema):
+    """Get an annotation with validation implemented for numbers and strings based on the schema.
+
+    :param annotation: an annotation from a field specification, as ``str``, ``ConstrainedStr`` or the
+      return value of ``constr(max_length=10)``
+    :param schema: an instance of Schema, possibly with declarations for validations and JSON Schema
+    :return: the same ``annotation`` if unmodified or a new annotation with validation in place
+    """
+    kwargs = {}
+    params_to_set = False
+    if (
+        isinstance(annotation, type)
+        and issubclass(annotation, (str,) + _numeric_types)
+        and not issubclass(annotation, bool)
+        and not any([issubclass(annotation, c) for c in _blacklist])
+    ):
+        if issubclass(annotation, str):
+            attrs = _str_attrs
+            kwargs = {'min_length': None, 'max_length': None, 'regex': None}
+            con = _map_types_const[str]
+        elif issubclass(annotation, _numeric_types):
+            attrs = _numeric_attrs
+            kwargs = {'gt': None, 'lt': None, 'ge': None, 'le': None}
+            for t in _numeric_types:
+                if issubclass(annotation, t):
+                    con = _map_types_const[t]
+                    break
+        for attr in attrs:
+            if hasattr(annotation, attr) and getattr(annotation, attr) is not None:
+                kwargs[attr] = getattr(annotation, attr)
+            elif hasattr(schema, attr) and getattr(schema, attr) is not None:
+                params_to_set = True
+                kwargs[attr] = getattr(schema, attr)
+        if params_to_set:
+            annotation = con(**kwargs)
+    return annotation
 
 
 class Field:
@@ -114,6 +211,7 @@ class Field:
             schema = Schema(value, **schema_from_config)
         schema.alias = schema.alias or schema_from_config.get('alias')
         required = value == Required
+        annotation = get_annotation_from_schema(annotation, schema)
         return cls(
             name=name,
             type_=annotation,