<!-- Thank you for your contribution! -->
<!-- Don't worry about making lots of commits on a pull request, they'll be squashed on merge anyway -->

## Change Summary

Add additional parameters to Schema for validation and annotation, in a single declaration.

Integrates and uses `constr`, `conint`, `confloat` and `condecimal`. Doesn't change any of the current functionality.

As a result, a current model created as:

```Python
from pydantic import BaseModel, constr, Schema

class Foo(BaseModel):
    a: constr(min_length=2, max_length=10) = Schema('bar')
```

can also be created as:

```Python
from pydantic import BaseModel, Schema

class Foo(BaseModel):
    a: str = Schema('bar', min_length=2, max_length=10)
```

by being able to declare `a` as a pure `str` makes it simpler for some editors to get the hint that the property `a` is a `str` and provide support for it (autocomplete, type error checking, etc).

### Note

The generated JSON Schema will output validation keywords of `minLength` and `maxLength` in this case. 

For numbers validations like `gt` it outputs `exclusiveMinimum`. 

Other option for the parameters would be to use the literal name used in JSON Schema, as `exclusiveMinimum`, but it wouldn't be very "Pythonic". 

And other option would be using JSON Schema names and change them to snake-case, as `exclusive_minimum`.

But I imagine the preferred option is to keep the same parameter names used in the rest of the code, as in `conint`, etc. Otherwise, let me know and I'll change it.

### Specific parameters added to `Schema`

* `description`
* `gt`
* `ge`
* `lt`
* `le`
* `min_length`
* `max_length`
* `regex`

<!-- Please give a short summary of the changes. -->

## Related issue number

I didn't create a specific issue for this. But the changes where discussed here: https://github.com/samuelcolvin/pydantic/issues/291#issuecomment-439056451

<!-- Are there any issues opened that will be resolved by merging this change? -->

## Performance Changes

pydantic cares about performance, if there's any risk performance changed on this PR, 
please run `make benchmark-pydantic` before and after the change:
* before:

```
                                pydantic time=0.775s, success=49.95%
                                pydantic time=0.794s, success=49.95%
                                pydantic time=0.792s, success=49.95%
                                pydantic time=0.773s, success=49.95%
                                pydantic time=0.790s, success=49.95%
                                pydantic best=0.773s, avg=0.785s, stdev=0.010s

                                pydantic best=25.770μs/iter avg=26.155μs/iter stdev=0.332μs/iter
```

* after:

```
                                pydantic time=0.794s, success=49.95%
                                pydantic time=0.801s, success=49.95%
                                pydantic time=0.809s, success=49.95%
                                pydantic time=0.765s, success=49.95%
                                pydantic time=0.797s, success=49.95%
                                pydantic best=0.765s, avg=0.793s, stdev=0.017s

                                pydantic best=25.507μs/iter avg=26.441μs/iter stdev=0.554μs/iter
```

## Checklist

* [x] Unit tests for the changes exist
* [x] Tests pass on CI and coverage remains at 100%
* [x] Documentation reflects the changes
* [x] No performance deterioration (if applicable)
* [x] `HISTORY.rst` has been updated
  * if this is the first change since a release, please add a new section
  * include the issue number or this pull request number `#<number>`
  * if you're not a regular contributer please include your github username `@whatever`