@tiangolo 2018-11-26
Thanks for the review! I'll come, check and fix in a couple days.