@tiangolo 2018-11-28
> I've changed the base for this PR so we can merge it without it going into master.
> 
> Fix the stuff that makes sense above, then I'll work on anything I want to change. Easier than going back and fourth for ever.

Great.

---

Do you want me to mark as "resolved" the comments in the review that are already done? 

I'm leaving them there so you can come back, check and confirm, before marking them as resolved. But let me know if you want me to mark the ones that are done and don't have a pending discussion.